sdk: correct missing casts

+++ bld/sdk/drwat/win/stat.c
@@ -330,7 +330,7 @@ INT_PTR CALLBACK SegMapDlgProc( HWND hwnd, UINT msg, WPARAM wparam, LPARAM lpara
                 str[4] = '\0';
                 seg = atoi( str );
                 if( DoGlobalEntryModule( &ge, DTTaskEntry.hModule, seg ) ) {
-                    DispMem( Instance, hwnd, ge.hBlock, (ge.dwSize == 1), 0, PMGetSelectorSize( ge.hBlock ) );
+                    DispMem( Instance, hwnd, (WORD)ge.hBlock, (ge.dwSize == 1), 0, PMGetSelectorSize( (WORD)ge.hBlock ) );
                 }
             }
             ret = true;
